fix: overhaul bugs across recording pipeline, player, storage, streaming and native layers
- Fix crash: waveform buffer used fixed-length List.filled with removeAt
- Fix num->double clamp assignments (setVolume, sliders, waveform bars)
- Fix AudioEngineService duplicated block & add channel-call timeouts
- Fix silent recording failure: native start now returns real success and
  errors propagate to UI (RECORD_START_FAILED), with writable-dir probe
- Fix fake format support: native capture now always stores playable WAV;
  UI warns that AAC/MP3/Opus export happens via share/zip
- Fix S3 SigV4: canonical query double-encoding, listObjectsV2 URL, client
  leaks; WebDAV: URL-encode segments and close http clients
- Add streaming connect-on-ready + heartbeat ping latency
- Add upload auto-retry with exponential backoff
- Guard invalid storage paths (probe write, fallback)
- CI: flutter create --no-overwrite so README/docs survive builds

+++ lib/ui/widgets/waveform_visualizer.dart
@@ -80,7 +80,7 @@ class WaveformVisualizer extends StatelessWidget {
               mainAxisAlignment: MainAxisAlignment.spaceEvenly,
               crossAxisAlignment: CrossAxisAlignment.center,
               children: waveform.map((amp) {
-                final barHeight = isRecording && !isPaused ? (amp * 65.0).clamp(4.0, 65.0) : 4.0;
+                final barHeight = isRecording && !isPaused ? (amp * 65.0).clamp(4.0, 65.0).toDouble() : 4.0;
                 return Container(
                   width: 3.5,
                   height: barHeight,
